Change the bootstrap fetch for movingAverage()
Originally this would fetch windowSize points before the original series
(bootstrap) and then try to create a new series merging the bootstrap
+ original series.
This caused problems when using the COARSE whisper feth strategy, since the
step of the bootstrap could be smaller than the step of the series.

To avoid step dances, re-fetch the series using a window of windowSize + timespan
of the original series and use this for calculation.